Restructure loadrecipe a bit
