+++ koocook_core/management/commands/loadrecipe.py
@@ -1,1 +1,12 @@
-from koocook_core.management.commands._loadrecipe import Command
+from django.core.management.base import BaseCommand
+
+
+class Command(BaseCommand):
+    help = 'load recipe data'
+
+    def handle(self, *args, **options):
+        self.stdout.write('calling read_data')
+        from ._add_path import add_datatrans
+        add_datatrans()
+        from ._loadrecipe.main import main
+        main()
